+++ src/azarrot/backends/backend_base.py
@@ -287,5 +287,7 @@ class BaseBackend(ABC):
         pass
 
     @abstractmethod
-    def generate_embeddings(self, request: EmbeddingsGenerationRequest) -> tuple[list[float], GenerationStatistics]:
+    def generate_embeddings(
+        self, request: EmbeddingsGenerationRequest
+    ) -> tuple[list[list[float]], GenerationStatistics]:
         pass