+++ tests/integration/openvino/test_qwen2.py
@@ -1,8 +1,9 @@
 from typing import Any
 
+import pytest
 from openai import OpenAI
 
-from azarrot.backends.ipex_llm_backend import BACKEND_ID_IPEX_LLM
+from azarrot.backends.openvino_backend import BACKEND_ID_OPENVINO
 from azarrot.models.model_manager import DEFAULT_MODEL_PRESETS
 from azarrot.server import Server
 from azarrot.tools import GLOBAL_TOOL_MANAGER
@@ -11,13 +12,13 @@ from azarrot.tools.tool import Tool, ToolDescription, ToolParameter
 QWEN2_CHAT_MODEL = "Qwen/Qwen2-7B-Instruct"
 
 
-def test_qwen2_hello(ipex_llm_server: Server) -> None:
-    ipex_llm_server.model_manager.load_huggingface_model(
-        QWEN2_CHAT_MODEL, BACKEND_ID_IPEX_LLM, "text-generation", skip_if_loaded=True
+def test_qwen2_hello(openvino_server: Server) -> None:
+    openvino_server.model_manager.load_huggingface_model(
+        QWEN2_CHAT_MODEL, BACKEND_ID_OPENVINO, "text-generation", skip_if_loaded=True
     )
 
     client = OpenAI(
-        base_url=f"http://{ipex_llm_server.config.host}:{ipex_llm_server.config.port}/v1", api_key="__TEST__"
+        base_url=f"http://{openvino_server.config.host}:{openvino_server.config.port}/v1", api_key="__TEST__"
     )
 
     completion = client.chat.completions.create(
@@ -29,16 +30,16 @@ def test_qwen2_hello(ipex_llm_server: Server) -> None:
     result = completion.choices[0].message
     assert result is not None
     assert result.content is not None
-    assert result.content.find("你好！有什么问题我可以帮助你解答吗？") >= 0
+    assert result.content.find("有什么我可以解答的问题") >= 0
 
 
-def test_qwen2_conversation(ipex_llm_server: Server) -> None:
-    ipex_llm_server.model_manager.load_huggingface_model(
-        QWEN2_CHAT_MODEL, BACKEND_ID_IPEX_LLM, "text-generation", skip_if_loaded=True
+def test_qwen2_conversation(openvino_server: Server) -> None:
+    openvino_server.model_manager.load_huggingface_model(
+        QWEN2_CHAT_MODEL, BACKEND_ID_OPENVINO, "text-generation", skip_if_loaded=True
     )
 
     client = OpenAI(
-        base_url=f"http://{ipex_llm_server.config.host}:{ipex_llm_server.config.port}/v1", api_key="__TEST__"
+        base_url=f"http://{openvino_server.config.host}:{openvino_server.config.port}/v1", api_key="__TEST__"
     )
 
     completion = client.chat.completions.create(
@@ -55,16 +56,17 @@ def test_qwen2_conversation(ipex_llm_server: Server) -> None:
     result = completion.choices[0].message
     assert result is not None
     assert result.content is not None
-    assert result.content.find("绿色对应的是数字2") >= 0
+    assert result.content.find("绿是2") >= 0
 
 
-def test_qwen2_tool_calling(ipex_llm_server: Server) -> None:
-    ipex_llm_server.model_manager.load_huggingface_model(
-        QWEN2_CHAT_MODEL, BACKEND_ID_IPEX_LLM, "text-generation", skip_if_loaded=True
+@pytest.mark.skip(reason="Qwen2 not stable on OpenVINO yet")
+def test_qwen2_tool_calling(openvino_server: Server) -> None:
+    openvino_server.model_manager.load_huggingface_model(
+        QWEN2_CHAT_MODEL, BACKEND_ID_OPENVINO, "text-generation", skip_if_loaded=True
     )
 
     client = OpenAI(
-        base_url=f"http://{ipex_llm_server.config.host}:{ipex_llm_server.config.port}/v1", api_key="__TEST__"
+        base_url=f"http://{openvino_server.config.host}:{openvino_server.config.port}/v1", api_key="__TEST__"
     )
 
     tools = [
@@ -149,10 +151,10 @@ class RRRTool(Tool):
         return kwargs["a"] + kwargs["b"] - 200
 
 
-def test_qwen2_internal_tool_calling(ipex_llm_server: Server) -> None:
-    ipex_llm_server.model_manager.load_huggingface_model(
+def test_qwen2_internal_tool_calling(openvino_server: Server) -> None:
+    openvino_server.model_manager.load_huggingface_model(
         QWEN2_CHAT_MODEL,
-        BACKEND_ID_IPEX_LLM,
+        BACKEND_ID_OPENVINO,
         "text-generation",
         skip_if_loaded=True,
         model_preset=DEFAULT_MODEL_PRESETS["qwen2"].with_enable_internal_tools(),
@@ -162,7 +164,7 @@ def test_qwen2_internal_tool_calling(ipex_llm_server: Server) -> None:
     GLOBAL_TOOL_MANAGER.register_tool(RRRTool())
 
     client = OpenAI(
-        base_url=f"http://{ipex_llm_server.config.host}:{ipex_llm_server.config.port}/v1", api_key="__TEST__"
+        base_url=f"http://{openvino_server.config.host}:{openvino_server.config.port}/v1", api_key="__TEST__"
     )
 
     completion = client.chat.completions.create(